1.0.6-snapshot

+++ Z-Bishop/src/main/java/com/isahl/chess/bishop/protocol/mqtt/model/QttProtocol.java
@@ -52,7 +52,7 @@ public abstract class QttProtocol
 
     public void setType(QttType type)
     {
-        mFrameHeader |= type.getValue();
+        mFrameHeader |= (byte) type.getValue();
     }
 
     @Override
